+++ client/views/reports/components/lastRunItem.js
@@ -1,0 +1,5 @@
+Template.lastRunItem.helpers({
+  isProblematic: function(){
+    return this.value.lastProb ? "text-danger" : "text-success";
+  }
+});